+++ misc/src/hkim0331/misc.clj
@@ -14,9 +14,7 @@
      (str/replace s re (str "$1" "...")))))
 
 ;; conversion between `char` and `int`
-;;
 ;; no use!
-;;
 ;; (defn char->int
 ;;   [c]
 ;;   (int c))
@@ -99,3 +97,30 @@
        (partition 2 1)
        (map #(<= (first %) (second %)))
        (every? true?)))
+
+;;; fold
+;; https://stackoverflow.com/questions/16800255/how-do-we-do-both-left-and-right-folds-in-clojure
+(defn fold-l
+  [f val coll]
+  (if (empty? coll)
+    val
+    (fold-l f (f val (first coll)) (rest coll))))
+
+(defn fold-r
+  [f val coll]
+  (if (empty? coll)
+    val
+    (f (fold-r f val (rest coll)) (first coll))))
+
+(comment
+  (fold-l + 0 (range 10))
+  (fold-r + 0 (range 10)))
+
+;;; inject
+(defn- inject-aux [coll ret]
+  (if (empty? coll)
+    ret
+    (inject-aux (rest coll) (conj ret (* (last ret) (first coll))))))
+
+(defn inject [coll]
+  (reduce + (inject-aux coll [1])))