+++ misc/src/hkim0331/tarai.clj
@@ -1,45 +1,16 @@
-(ns hkim0331.misc.tarai)
+(ns hkim0331.tarai)
+
+(def ^:private version "0.0.1")
 
 (defn tarai
- [x y z]
- (if (<= x y)
-   y
-   (tarai (tarai (dec x) y z)
-          (tarai (dec y) z x)
-          (tarai (dec z) x y))))
+  [x y z]
+  (if (<= x y)
+    y
+    (tarai (tarai (dec x) y z)
+           (tarai (dec y) z x)
+           (tarai (dec z) x y))))
 
-
-;; nuc で実験。コンテナはひどく遅い。
-(comment
-  (time (tarai 15 7 0)))
-;; macos "Elapsed time: 17523.399916 msecs"
-;;       "Elapsed time: 28853.187034 msecs"
-;; docker-compose up
-;;       "Elapsed time: 193561.018473 msecs"
-
-(comment
-  (time (tarai 15 7 0)))
-;; m2:
-;;  "Elapsed time: 12051.361667 msecs"
-;;  "Elapsed time: 12050.578041 msecs"
-;;  "Elapsed time: 12041.758583 msecs"
-;; nuc:
-;;  "Elapsed time: 34066.433117 msecs"
-
-;;(time (tarai 14 7 0))
-;; clj : 1883msec
-;; cljs: 2665msec
-
-;; openjdk@11
-;;(time (my.tarai/tarai 14 7 0))
-;;"Elapsed time: 1665.51025 msecs"
-;;14
-
-;; docker compose
-;; (time (my.tarai/tarai 14 7 0))
-;; "Elapsed time: 1671.123333 msecs"
-;; 14
-
+;; 戻り値が違う tak 関数
 (defn tak [x y z]
   (if (<= x y)
     z
@@ -47,7 +18,21 @@
          (tak (dec y) z x)
          (tak (dec z) x y))))
 
-;;(time (tak 12 6 0))
+
+;; nuc で実験。コンテナはひどく遅い。
+(comment
+  (time (tarai 15 7 0)))
+;; m3    "Elapsed time: 10744.901625 msecs"
+;; macos "Elapsed time: 17523.399916 msecs"
+;;       "Elapsed time: 28853.187034 msecs"
+;; docker-compose up
+;;       "Elapsed time: 193561.018473 msecs"
+;; m2:
+;;  "Elapsed time: 12051.361667 msecs"
+;;  "Elapsed time: 12050.578041 msecs"
+;;  "Elapsed time: 12041.758583 msecs"
+;; nuc:
+;;  "Elapsed time: 34066.433117 msecs"
 
 ;; memoized
 (defn memoized-tarai [x y z]
@@ -59,11 +44,12 @@
                     (let [result (tarai (tarai (- x 1) y z)
                                         (tarai (- y 1) z x)
                                         (tarai (- z 1) x y))]
-                      (do (swap! memo assoc [x y z] result)
-                          result)))))]
+                      (swap! memo assoc [x y z] result)
+                      result))))]
       (tarai x y z))))
 
-;;(time (memoized-tarai 14 7 0))
+(time (memoized-tarai 15 7 0))
+;; "Elapsed time: 1.675791 msecs"
 
 ;; lazy function
 (defn lazy-tarai [x y z]
@@ -75,8 +61,8 @@
                      (fn [] (tarai (fn [] (- (fz) 1)) fx fy)))))]
     (tarai (fn [] x) (fn [] y) (fn [] z))))
 
-(comment
-  (time (lazy-tarai 14 7 0)))
 
+(time (lazy-tarai 15 7 0))
+;; "Elapsed time: 1.012375 msecs"