+++ misc/test/hkim0331/misc_test.clj
@@ -1,6 +1,6 @@
 (ns hkim0331.misc-test
   (:require
-   [clojure.test :refer :all]
+   [clojure.test :refer [deftest testing is]]
    [hkim0331.misc :refer :all]))
 
 (deftest abbrev-test
@@ -14,3 +14,12 @@
     (is (= [] (qsort [])))
     (is (= (range 10) (qsort (reverse (range 10)))))
     (is (= true (ordered? (range 10))))))
+
+(deftest compress-test
+  (testing "test compress/expand"
+    (let [s1 ""
+          s2 "aaaaaaaaabbbbbbbdacccccccc"
+          s3 "abcabcdef"]
+      (is (= s1 (expand (compress s1))))
+      (is (= s2 (expand (compress s2))))
+      (is (= s3 (expand (compress s3)))))))